1.修复高版本评论优先显示最热失效 2.修改部分int类型为long类型减少崩溃 3.修改日志输出模式，修复NDK内存泄漏问题 4.合并分支，兼容极速版 5.新增音质优先还是匹配度优先选项

+++ app/src/main/java/com/raincat/dolby_beta/model/UserInfoBean.java
@@ -33,15 +33,15 @@ public class UserInfoBean {
     }
 
     public static class ProfileBean {
-        private int userId = -1;
+        private long userId = -1;
         private int userType = 0;
         private String nickname = "";
 
-        public int getUserId() {
+        public long getUserId() {
             return userId;
         }
 
-        public void setUserId(int userId) {
+        public void setUserId(long userId) {
             this.userId = userId;
         }
